+++ src/components/spartan/SDropdown/types.ts
@@ -1,0 +1,10 @@
+import { type FunctionalComponent } from 'vue';
+
+export type TDropdownProps = {
+    leftToRight: boolean;
+};
+
+export type TDropdownItemProps = {
+    disabled: boolean;
+    icon: FunctionalComponent;
+}