+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- `TDropdownItemProps` now exports his types.
+- `TDropdownProps` now exports his types.
 - `TDefinitionTerm` now exports his types.
 - `SButtonGroupItem` now exports his types.
 - `SButton` now exports his types.